[MIST-581] Implement GroupMetricHandler option 1 (#595)
* [MIST-581] Implement GroupMetricHandler option 1

* Amendment on comments

* [MIST-581] address comments

+++ src/main/java/edu/snu/mist/core/task/eventProcessors/EventProcessorManager.java
@@ -37,5 +37,5 @@ public interface EventProcessorManager extends AutoCloseable {
    * It will delete existing event processors if the current # of event processors > adjustNum.
    * @param adjustNum the number of event processors to be adjusted.
    */
-  void adjustEventProcessorNum(int adjustNum);
+  void adjustEventProcessorNum(long adjustNum);
 }
